seventh commit

+++ application/src/main/java/com/blessy/application/repository/ContinentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ContinentRepository extends JpaRepository<Continent, Integer> {
+public interface ContinentRepository extends JpaRepository<Continent, Long> {
 
     Optional<Continent> findByName(String name);
 }
